create competent structure
