+++ src/common/definitions.h
