+++ src/common/logger.c
@@ -1,5 +1,5 @@
-#include "logger.h"
-#include "util.h"
+#include "common/logger.h"
+#include "common/util.h"
 
 #include <stdlib.h>
 #include <stdio.h>
