+++ src/common/pe32.h
