+++ src/inject/crt.h
@@ -1,7 +1,7 @@
 #ifndef   CREATE_REMOTE_THREAD_HEADER
 #define   CREATE_REMOTE_THREAD_HEADER
 
-#include "definitions.h"
+#include "common/definitions.h"
 
 enum status_t create_remote_thread(const struct parsed_args_t*);
