+++ src/inject/loadlibrary.h
@@ -1,7 +1,7 @@
 #ifndef   LOAD_LIBRARY_HEADER
 #define   LOAD_LIBRARY_HEADER
 
-#include  "definitions.h"
+#include  "common/definitions.h"
 
 enum status_t load_library_a(const struct parsed_args_t*);
 enum status_t load_library_w(const struct parsed_args_t*);
