+++ src/pidjeon/parser.h
@@ -1,7 +1,7 @@
 #ifndef   PARSER_HEADER
 #define   PARSER_HEADER
 
-#include  "definitions.h"
+#include  "common/definitions.h"
 
 enum errcode_t parse_arguments(int argc, char** argv, struct parsed_args_t* parsed_args);
